% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{add_juxtacrine_view}
\alias{add_juxtacrine_view}
\title{Title}
\usage{
add_juxtacrine_view(current.views, positions, neighbor.thr = 15)
}
\arguments{
\item{current.views}{}

\item{positions}{}

\item{neighbor.thr}{}
}
\value{

}
\description{
Title
}
