% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{add_juxtaview}
\alias{add_juxtaview}
\title{Add juxtaview to the pipeline}
\usage{
add_juxtaview(
  current.views,
  positions,
  neighbor.thr = 15,
  cached = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{current.views}{}

\item{positions}{}

\item{neighbor.thr}{}

\item{cached}{}

\item{verbose}{}
}
\value{

}
\description{
Add juxtaview to the pipeline
}
\examples{
# TBD
}
