% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{add_paracrine_view}
\alias{add_paracrine_view}
\title{Title}
\usage{
add_paracrine_view(current.views, positions, l, approx = 1, ncells = NULL)
}
\arguments{
\item{current.views}{}

\item{positions}{}

\item{l}{}

\item{approx}{}

\item{ncells}{}
}
\value{

}
\description{
Title
}
