% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{add_paraview}
\alias{add_paraview}
\title{Add paraview to the pipeline}
\usage{
add_paraview(
  current.views,
  positions,
  l,
  approx = 1,
  ncells = NULL,
  cached = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{current.views}{}

\item{positions}{}

\item{l}{}

\item{approx}{}

\item{ncells}{}

\item{cached}{}

\item{verbose}{}
}
\value{

}
\description{
Add paraview to the pipeline
}
