% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aggregate_results_subset}
\alias{aggregate_results_subset}
\title{Aggregate a subset of results}
\usage{
aggregate_results_subset(misty.results, folders)
}
\arguments{
\item{misty.results}{a results list generated by
\code{\link[MISTy:collect_results]{collect_results()}}.}

\item{folders}{Paths to folders containing the raw results from
\code{\link[MISTy:run_misty]{run_misty()}}.}
}
\value{
the \code{misty.results} list with an added list item
    \code{importances.aggregated.subset} containing the aggregated importances
    for the subset of \code{folders}.
}
\description{
Aggregate a subset of results
}
\seealso{
\code{\link[MISTy:collect_results]{collect_results()}} to generate a
    results list from raw results.
}
