% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-functions.R
\name{bagged_mars_model}
\alias{bagged_mars_model}
\title{Bagged MARS}
\usage{
bagged_mars_model(view_data, target, seed, n.bags = 50, ...)
}
\arguments{
\item{view_data}{\code{tibble} containing the expression of the markers
for each cell (rows = cells, cols = markers)}

\item{target}{\code{string} indicating the target marker, passed by run_misty()}

\item{seed}{\code{integer} passed by run_misty()}

\item{n.bags}{\code{integer} indicating number of bags
(bootstrap aggregation samples) used for bagging}

\item{...}{all additional parameters are passed to the chosen ML model}
}
\description{
Function that can be passed to \code{run_misty()} via \code{run_misty(views, 
model.function = bagged_mars_model)} to model each view using bagged MARS,
meaning multivariate adaptive spline regression models trained with 
bootstrap aggregation samples. The algorithm is implemented in the \code{earth} 
R package.
}
\details{
The following parameters are the default configuration: \code{degree = 2}.
Furthermore 50 base learners are used by default (see \code{n.bags}).
}
