% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{build_model}
\alias{build_model}
\title{Train a multi-view model for a single target}
\usage{
build_model(views, target, seed = 42, cv.folds = 10, cached = TRUE, ...)
}
\arguments{
\item{views}{view composition.}

\item{target}{name of the target marker to train models for.}

\item{seed}{seed used for random sampling to ensure reproducibility.}

\item{cv.folds}{number of cross-validation folds to consider for estimating
the performance of the multi-view models.}

\item{cached}{a \code{logical} indicating whether to cache the trained models
and to reuse previously cached ones if they already exist for this sample.}

\item{...}{all additional parameters are passed to
\code{\link[ranger:ranger]{ranger()}} for training the view-specific models
(see Details for defaults).}
}
\value{
A list containing the trained meta-model, a list of trained
view-specific models and performance estimates.
}
\description{
Trains individual Random Forest models for each view, a linear meta-model
using the out-of-bag predictions of the view specific models and estimates
the overall performance by cross-validation.
}
\details{
Default values passed to \code{\link[ranger:ranger]{ranger()}} for training the
view-specific models: \code{num.trees = 100}, \code{importance = "impurity"},
\code{num.threads = 1}, \code{seed = seed}.
}
