% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clear_cache}
\alias{clear_cache}
\title{Clear cached objects}
\usage{
clear_cache(id = NULL)
}
\arguments{
\item{id}{the unique id of the sample.}
}
\value{
None (\code{NULL})
}
\description{
Purge the cache or clear the cached objects for a single sample.
}
\details{
The cached objects are removed from disk and cannot be retrieved. Whenever
possible specifying an \code{id} is reccomended. If \code{id = NULL} all
contents of the folder \file{.misty.temp} will be removed.
}
