% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collect_results}
\alias{collect_results}
\title{Collect and aggregate raw results}
\usage{
collect_results(folders)
}
\arguments{
\item{folders}{Paths to folders containing the raw results from \code{\link{run_misty}}.}
}
\value{
TODO: return value
}
\description{
Collect and aggregate raw results
}
\seealso{
\code{\link{run_misty}} to train models and generate results.
}
