% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{create_view}
\alias{create_view}
\title{Create a view from a data frame}
\usage{
create_view(name, data, abbrev = name)
}
\arguments{
\item{name}{}

\item{data}{}

\item{abbrev}{}
}
\value{

}
\description{
Create a view from a data frame
}
\examples{
#TBD
}
