% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-functions.R
\name{linear_model}
\alias{linear_model}
\title{Simple Linear Model}
\usage{
linear_model(view_data, target, seed, k = 10, ...)
}
\arguments{
\item{view_data}{\code{tibble} containing the expression of the markers for
each cell (rows = cells, cols = markers)}

\item{target}{\code{string} indicating the target marker, passed by run_misty()}

\item{seed}{\code{integer} passed by run_misty()}

\item{k}{\code{integer} indicating the folds used in cross validation to compute the
unbiased estimates}

\item{...}{all additional parameters are passed to the chosen ML model}
}
\description{
Function that can be passed to \code{run_misty()} via \code{run_misty(views, 
model.function = linear_model)} to model each view using a simple linear model
}
