% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-functions.R
\name{mars_model}
\alias{mars_model}
\title{Bagged MARS}
\usage{
mars_model(view_data, target, seed, approx = 1, k = 10, ...)
}
\arguments{
\item{view_data}{\code{tibble} containing the expression of the markers for
each cell (rows = cells, cols = markers)}

\item{target}{\code{string} indicating the target marker, passed by run_misty()}

\item{seed}{\code{integer} passed by run_misty()}

\item{approx}{\code{double} indicating the fraction of the training data used
for training in each fold}

\item{k}{\code{integer} indicating the folds used in cross validation to
compute the unbiased estimates}

\item{...}{all additional parameters are passed to the chosen ML model}
}
\description{
Function that can be passed to \code{run_misty()} via \code{run_misty(views, 
model.function = mars_model)} to model each view using a multivariate adaptive
spline regression model. The algorithm is implemented in the \code{earth} R package.
}
\details{
The following parameters are the default configuration: \code{degree = 2}
}
