% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{merge_two}
\alias{merge_two}
\title{Function to merge named arguments from two lists without removing NULL entries}
\usage{
merge_two(l1, l2)
}
\arguments{
\item{l1}{list 1}

\item{l2}{list 2}
}
\description{
Function to merge named arguments from two lists without removing NULL entries
}
