% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_contrast_results}
\alias{plot_contrast_results}
\title{Title}
\usage{
plot_contrast_results(
  misty.results.from,
  misty.results.to,
  views = NULL,
  cutoff.from = 1,
  cutoff.to = 1
)
}
\arguments{
\item{misty.results.from}{Results from}

\item{misty.results.to}{Results to}

\item{views}{Views vector}

\item{cutoff.from}{Cutoff from}

\item{cutoff.to}{Cutoff to}
}
\value{
The \code{misty.results} list (invisibly).
}
\description{
Description
}
\seealso{
\code{\link{collect_results()}} to generate a results list from raw 
    results.

Other plotting functions: 
\code{\link{plot_contrast_heatmap}()},
\code{\link{plot_improvement_stats}()},
\code{\link{plot_interaction_communities}()},
\code{\link{plot_interaction_heatmap}()},
\code{\link{plot_view_contributions}()}
}
\concept{plotting functions}
