% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_improvement_stats}
\alias{plot_improvement_stats}
\title{Plot observed performance and improvement per target}
\usage{
plot_improvement_stats(misty.results, measure = "gain.R2")
}
\arguments{
\item{measure}{}
}
\value{

}
\description{
Plot observed performance and improvement per target
}
\examples{
# TBD
}
