% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_interaction_communities}
\alias{plot_interaction_communities}
\title{Plot marker interaction communities}
\usage{
plot_interaction_communities(
  misty.results,
  view,
  cutoff = 1,
  trim = 0,
  trim.measure = c("gain.R2", "multi.R2", "intra.R2", "gain.RMSE", "multi.RMSE",
    "intra.RMSE"),
  resolution = 1,
  path = NULL
)
}
\arguments{
\item{misty.results}{a results list generated by
\code{\link{collect_results}()}.}

\item{view}{abbreviated name of the view.}

\item{cutoff}{importance threshold. Importances below this value will
be colored white in the heatmap and considered as not relevant.}

\item{trim}{display targets with performance value above (if R2 or gain) or
below (otherwise) this value only.}

\item{trim.measure}{the measure used for trimming.}

\item{resolution}{The resolution parameter for the Leiden community detection
based clustering. The higher the value the higher the number of communities
with smaller number of members.}

\item{path}{path to write a graphml file of the resulting graph}
}
\value{
The \code{misty.results} list (invisibly).
}
\description{
Identify and plot a graph of marker interaction communities.
}
\details{
The communities are identified using the Louvain algorithm. Communities can
be extracted only from views that have the same predictor and target markers.
}
\examples{

misty.results <- collect_results("example.sqm")

misty.results \%>\%
  plot_interaction_communities("intra") \%>\%
  plot_interaction_communities("para.10")

misty.results \%>\%
  plot_interaction_communities("para.10", cutoff = 0.5)
}
\seealso{
\code{\link{collect_results}()} to generate a
    results list from raw results.

Other plotting functions: 
\code{\link{plot_contrast_heatmap}()},
\code{\link{plot_contrast_results}()},
\code{\link{plot_improvement_stats}()},
\code{\link{plot_interaction_heatmap}()},
\code{\link{plot_view_contributions}()}
}
\concept{plotting functions}
