% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-functions.R
\name{random_forest_model}
\alias{random_forest_model}
\title{Random Forest Implementation}
\usage{
random_forest_model(view_data, target, seed, ...)
}
\arguments{
\item{view_data}{\code{tibble} containing the expression of the markers for
each cell (rows = cells, cols = markers)}

\item{target}{\code{string} indicating the target marker, passed by run_misty()}

\item{seed}{\code{integer} passed by run_misty()}

\item{...}{all additional parameters are passed to the chosen ML model}
}
\description{
The default function to model each view based on random forest. The 
algorithm is implemented in the \code{ranger} R package.
}
\details{
The following parameters are the default configuration: \code{num.trees = 100},
 \code{importance = "impurity"}, \code{num.threads = 1}, \code{seed = seed}.

Can be explicitly called via 
\code{run_misty(views, model.function = random_forest_model)}
}
