% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misty.R
\name{run_sliding_misty}
\alias{run_sliding_misty}
\title{Train sliding MISTy models}
\usage{
run_sliding_misty(
  views,
  positions,
  window,
  overlap = 50,
  results.folder = "results",
  minu = 100,
  ...
)
}
\arguments{
\item{views}{view composition.}

\item{positions}{a \code{data.frame}, \code{tibble} or a \code{matrix}
with named coordinates in columns and rows for each spatial unit ordered
as in the intraview.}

\item{window}{size of the}

\item{overlap}{overlap of consecutive windows (percentage).}

\item{results.folder}{path to the top level folder to store raw results.}

\item{minu}{minimum number of spatial units in the window.}

\item{...}{all other parameters are passed to \code{\link{run_misty}()}.}
}
\value{
Path to the result folder(s) that can be passed to
    \code{\link{collect_results}()}.
}
\description{
Train local MISTy models by sliding a window across the sample as captured by the
view composition.
}
\examples{
# TODO

}
