% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{svm_model}
\alias{svm_model}
\title{SVM Implementation}
\usage{
svm_model(view_data, target, seed, approx = 0.4, k = 10, ...)
}
\arguments{
\item{view_data}{\code{tibble} containing the expression of the markers for each cell
(rows = cells, cols = markers)}

\item{target}{\code{string} indicating the target marker, passed by run_misty()}

\item{seed}{\code{integer} passed by run_misty()}

\item{k}{\code{integer} indicating the folds used in cross validation to compute the
unbiased estimates}

\item{...}{all additional parameters are passed to the chosen ML model}
}
\description{
Function that can be passed to \code{run_misty()} via \code{run_misty(views, 
model.function = gradient_boosting_model)} to model each view using a linear
support vector machine. The algorithm is implemented in the \code{kernlab} R package.
}
\details{
The following parameters are the default configuration: \code{kernel = "vanilladot"} 
(linear kernel), \code{C = 1}, \code{type = "eps-svr"}.
}
