% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{add_views}
\alias{add_views}
\title{Add custom views to the current pipeline}
\usage{
add_views(current.views, new.views)
}
\arguments{
\item{new.views}{}
}
\value{

}
\description{
Add custom views to the current pipeline
}
