% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collect_results}
\alias{collect_results}
\title{Collect and aggregate MISTy results}
\usage{
collect_results(folders)
}
\arguments{
\item{folders}{}
}
\value{

}
\description{
Collect and aggregate MISTy results
}
\examples{
# TBD
}
