% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{create_initial_view}
\alias{create_initial_view}
\title{Create the basic intrinsic view for MISTy}
\usage{
create_initial_view(table, unique.id = NULL)
}
\arguments{
\item{table}{}

\item{unique.id}{}
}
\value{

}
\description{
Create the basic intrinsic view for MISTy
}
\examples{
#TBD
}
