% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_contrast_heatmap}
\alias{plot_contrast_heatmap}
\title{Title}
\usage{
plot_contrast_heatmap(misty.results, from.view, to.view, cutoff = 1)
}
\arguments{
\item{cutoff}{}
}
\value{

}
\description{
Title
}
\examples{
# TBD
}
