% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_contrast_results}
\alias{plot_contrast_results}
\title{Plot heatmap of contrast between two result lists}
\usage{
plot_contrast_results(
  misty.results.from,
  misty.results.to,
  views = NULL,
  cutoff.from = 1,
  cutoff.to = 1
)
}
\arguments{
\item{misty.results.from, misty.results.to}{a results list generated by
\code{\link[MISTy:collect_results]{collect_results()}}.}

\item{views}{one or more abbreviated names of views.}

\item{cutoff.from, cutoff.to}{importance thresholds respective to the result
lists.}
}
\value{
The \code{misty.results.from} list (invisibly).
}
\description{
Plot interexperiment contrast of views. The heatmaps show the interactions
that are present and have importance above a \code{cutoff.to} value in the
\code{views} of \code{misty.results.to} but not present or have importance
below \code{cutoff.from} in the \code{views} of \code{misty.results.from}.
}
\seealso{
\code{\link[MISTy:collect_results]{collect_results()}} to generate a
    results list from raw results.

Other plotting functions: 
\code{\link{plot_contrast_heatmap}()},
\code{\link{plot_improvement_stats}()},
\code{\link{plot_interaction_communities}()},
\code{\link{plot_interaction_heatmap}()},
\code{\link{plot_view_contributions}()}
}
\concept{plotting functions}
