% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_interaction_communities}
\alias{plot_interaction_communities}
\title{Title}
\usage{
plot_interaction_communities(misty.results, view, cutoff = 1)
}
\arguments{
\item{cutoff}{}
}
\value{

}
\description{
Title
}
