% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_view_contributions}
\alias{plot_view_contributions}
\title{Title}
\usage{
plot_view_contributions(misty.results)
}
\arguments{
\item{misty.results}{}
}
\value{

}
\description{
Title
}
\examples{
# TBD
}
