% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{remove_views}
\alias{remove_views}
\title{Remove views from the pipeline}
\usage{
remove_views(current.views, view.names)
}
\arguments{
\item{view.names}{}
}
\value{

}
\description{
Remove views from the pipeline
}
