% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misty.R
\name{run_misty}
\alias{run_misty}
\title{Run MISTy}
\usage{
run_misty(
  views,
  results.folder = "results",
  seed = 42,
  target.subset = NULL,
  cv.folds = 10,
  cached = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Run MISTy
}
