% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misty.R
\name{run_misty}
\alias{run_misty}
\title{Train MISTy models}
\usage{
run_misty(
  views,
  results.folder = "results",
  seed = 42,
  target.subset = NULL,
  cv.folds = 10,
  cached = FALSE,
  append = FALSE,
  method = "bag",
  learner = "ranger",
  n.vars = NULL,
  n.learners = 100,
  bypass.intra = FALSE,
  ...
)
}
\arguments{
\item{views}{view composition.}

\item{results.folder}{path to the top level folder to store raw results.}

\item{seed}{seed used for random sampling to ensure reproducibility.}

\item{target.subset}{subset of targets to train models for. If \code{NULL},
models will be trained for markers in the intraview.}

\item{cv.folds}{number of cross-validation folds to consider for estimating
the performance of the multi-view models and the view-specific models
if \code{method = "cv"}}

\item{cached}{a \code{logical} indicating whether to cache the trained models
and to reuse previously cached ones if they already exist for this sample.}

\item{append}{a \code{logical} indicating whether to append the performance
and coefficient files in the \code{results.folder}. Consider setting to
\code{TRUE} when rerunning a workflow with different \code{target.subset}
parameters.}

\item{method}{a string indicating whether to use bagging "bag" or cross-
valdiation to train the view-specific models and to get unbiased predictions
for the linear meta model}

\item{learner}{a string indicating which ML model to use to model the 
views, possible values are "ranger" for random forest, "lm" for linear
model, "linearSVM" for support vector machine with linear kernel, and
"earth" for multivariate adaptive regression splines.}

\item{n.vars}{number indicating how many variables should be used for 
training the view-specific models if \code{method = "bag"}. For each base
learner a sample of size \code{n.vars} from the predictors is taken. 
If \code{learner = "ranger"} \code{n.vars} corresponds to \code{mtry}, 
meaning the number of variables that are considered at each split.}

\item{n.learners}{number indicating how many learners should be used for 
the bagging model, only relevant if \code{method = "bag"}}

\item{bypass.intra}{a \code{logical} indicating whether to train a baseline
model using the intraview data (see Details).}

\item{...}{all additional parameters are passed to the chosen ML model for
training the view-specific models}
}
\value{
Path to the results folder that can be passed to
    \code{\link{collect_results}()}.
}
\description{
Trains multi-view models for all target markers, estimates the performance,
the contributions of the view specific models and the importance of predictor
markers for each target marker.
}
\details{
If \code{bypass.intra} is set to \code{TRUE} all variable in the intraview
the intraview data will be treated as targets only. The baseline intraview
model in this case is a trivial model that predicts the average of each
target. If the intraview has only one variable this switch is automatically
set to \code{TRUE}.

Default ML model to train the view-specific views is 
\code{\link[ranger]{ranger}()}. The following parameters are the default
configuration: \code{num.trees = 100}, \code{importance = "impurity"},
\code{num.threads = 1}, \code{seed = seed}.
}
\examples{
# Create a view composition of an intraview and a paraview with radius 10 then
# run MISTy for a single sample.

library(dplyr)

# get the expression data
data("synthetic")
expr <- synthetic[[1]] \%>\% select(-c(row, col, type))
# get the coordinates for each cell
pos <- synthetic[[1]] \%>\% select(row, col)

# compose
misty.views <- create_initial_view(expr) \%>\% add_paraview(pos, l = 10)

# run with default parameters
run_misty(misty.views)
}
\seealso{
\code{\link{create_initial_view}()} for
    starting a view composition.
}
