% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misty.R
\name{run_misty}
\alias{run_misty}
\title{Train MISTy models}
\usage{
run_misty(
  views,
  results.folder = "results",
  seed = 42,
  target.subset = NULL,
  cv.folds = 10,
  cached = TRUE,
  ...
)
}
\arguments{
\item{views}{view composition.}

\item{results.folder}{path to the top level folder to store raw results.}

\item{seed}{seed used for random sampling to ensure reproducibility.}

\item{target.subset}{subset of targets to train models for. If \code{NULL},
models will be trained for markers in the intraview.}

\item{cv.folds}{number of cross-validation folds to consider for estimating
the performance of the multi-view models.}

\item{cached}{a \code{logical} indicating whether to cache the trained models
and to reuse previously cached ones if they already exist for this sample.}

\item{...}{all additional parameters are passed to
\code{\link[ranger]{ranger}()} for training the view-specific models
(see Details for defaults).}
}
\value{
Path to the results folder that can be passed to
    \code{\link{collect_results}()}.
}
\description{
Trains multi-view models for all target markers, estimates the performance,
the contributions of the view specific models and the importance of predictor
markers for each target marker.
}
\details{
Default values passed to \code{\link[ranger]{ranger}()} for training the
view-specific models: \code{num.trees = 100}, \code{importance = "impurity"},
\code{num.threads = 1}, \code{seed = seed}.
}
\seealso{
\code{\link{create_initial_view}()} for
    starting a view composition.
}
